Brought UserGroupSuggestOracle into it's own file.
