+++ src/main/java/org/sagebionetworks/web/client/widget/search/UserGroupSuggestBoxView.java
@@ -2,8 +2,8 @@ package org.sagebionetworks.web.client.widget.search;
 
 import org.sagebionetworks.repo.model.UserGroupHeaderResponsePage;
 import org.sagebionetworks.web.client.SynapseView;
-import org.sagebionetworks.web.client.widget.search.UserGroupSuggestBox.UserGroupSuggestOracle;
-import org.sagebionetworks.web.client.widget.search.UserGroupSuggestBox.UserGroupSuggestOracle.UserGroupSuggestion;
+import org.sagebionetworks.web.client.widget.search.UserGroupSuggestOracle;
+import org.sagebionetworks.web.client.widget.search.UserGroupSuggestOracle.UserGroupSuggestion;
 
 import com.google.gwt.user.client.ui.IsWidget;
